refactor: Clarify/improve structure for registration
chore: add specific version number of public builds
